Define Config struct for YAML configuration
A struct in Go is like a class in other languages, but simpler - it's
just a collection of fields. Here we define a Config struct with a
Words field that will hold a slice (array) of strings.

The yaml tag tells the YAML parser how to map YAML fields to Go struct
fields. This is called 'struct tags' in Go.